[feat] Consider the sampling time for expensive sampler
The previous version could behave like the last sampling starts before
the second last sampling ends. But that is not the case for the most
samplers, so I changed so that I will consider the second last sample
end timing and the latest sampling's timer starts after the second last
sample ends.