Change timehash calculation slightly
As I changed the multiple worker instantiations to non-multiprocessing,
now threadID is always identical unless we use different processes as
in NePS. However, in that case, os.getpid() is sufficient.
So I removed threading.get_ident() and added np.random.random() instead.
By doing so, it is more likely to yield unique timehash.